Make numbers table as a single module.
