+++ src/query/service/src/table_functions/numbers/mod.rs
@@ -1,4 +1,4 @@
-//  Copyright 2021 Datafuse Labs.
+//  Copyright 2022 Datafuse Labs.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -12,15 +12,9 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-mod async_crash_me;
-mod numbers;
-mod sync_crash_me;
-mod table_function;
-mod table_function_factory;
+mod numbers_part;
+mod numbers_table;
 
-pub use numbers::generate_numbers_parts;
-pub use numbers::NumbersPartInfo;
-pub use numbers::NumbersTable;
-pub use table_function::TableFunction;
-pub use table_function_factory::TableArgs;
-pub use table_function_factory::TableFunctionFactory;
+pub use numbers_part::generate_numbers_parts;
+pub use numbers_part::NumbersPartInfo;
+pub use numbers_table::NumbersTable;
