+++ src/query/service/tests/it/table_functions/mod.rs
@@ -12,5 +12,4 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.W
 
-mod memory_block_part;
 mod numbers_table;
